refactor mcp proxy client loop
